requirements.txt
fastapi==0.110.0
h11==0.14.0
pydantic~=2.7.3
typing-extensions
uvicorn[standard]~=0.30.1
python-multipart~=0.0.16
python-keycloak==3.12.0
python-dotenv~=1.0.1
fastapi-pagination

# database
sqlmodel~=0.0.19
asyncpg~=0.29.0
psycopg2-binary
psycopg2
SQLAlchemy~=2.0.30
fastapi-sqlalchemy
# alembic
sqlalchemy-utils
pytz~=2024.1

numpy
pandas
tf-keras==2.17.0

requests~=2.32.3
starlette~=0.36.3
typing_extensions~=4.12.2
cryptography~=42.0.8

pytest
pytest-cov
httpx
pytest-asyncio
limits

Pillow

openpyxl
accelerate>=0.26.0

selenium
groq
webdriver_manager
openai